requirements.txt
# --- LangChain Core & Partners ---
# LangChain v1.0 released on Oct 17, 2025
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langchain-anthropic>=1.0.0,<2.0.0
langchain-google-genai>=3.0.0,<5.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=1.0.0,<2.0.0

# --- LLM Providers ---
openai>=1.0.0
anthropic>=0.34.0
google-generativeai>=0.8.0

# --- Vector DBs & Tools ---
chromadb>=0.5.0
faiss-cpu>=1.8.0
pymed>=0.8.9
tiktoken>=0.7.0
lark>=1.1.0

# --- Utilities ---
python-dotenv>=1.0.0
apscheduler